rename and cleaning up
