+++ .gitignore
@@ -48,6 +48,9 @@ Thumbs.db
 # package files
 
 *egg-info*
+dist/
+build/
+*.dist-info/
 
 #docs
