+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: HUME - Mixed Graph Estimation
+site_name: HUMPY - Mixed Graph Estimation
 
 repo_url: https://github.com/konstantingoe/mixed-gm
 site_url: https://konstantingoe.github.io/mixed-gm/
