+++ pyproject.toml
@@ -1,27 +1,54 @@
 [build-system]
-requires = ["hatchling"]
+requires = ["hatchling >= 1.26"]
 build-backend = "hatchling.build"
 
 [project]
-name = "hume"
+name = "humpy"
 requires-python = ">=3.10.0"
-authors = [{ name = "Konstantin Göbler", email = "konstantin.goebler@tum.de" }]
-maintainers = [
-    { name = "Konstantin Göbler", email = "konstantin.goebler@tum.de" },
-    { name = "Stephan Haug", email = "stephan.haug@tum.de" },
+authors = [
+    { name = "Konstantin Göbler", email = "konstantingoebler@gmail.com" },
 ]
-license = { file = "LICENSE" }
-description = "High-dimensional Undirected Mixed graph Estimation - Python implementation"
+maintainers = [
+    { name = "Konstantin Göbler", email = "konstantingoebler@gmail.com" },
+]
+license = "MIT"
+license-files = ["LICENSE"]
+description = "High-dimensional Undirected Mixed graph estimation in PYthon"
+keywords = [
+    "graphical models",
+    "mixed data",
+    "graphical lasso",
+    "undirected graphs",
+    "statistics",
+    "Gaussian copula",
+]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Topic :: Scientific/Engineering :: Mathematics",
+    "Topic :: Scientific/Engineering :: Information Analysis",
+]
 dependencies = [
     "numpy>=1.21.0",
     "scipy>=1.7.0",
     "scikit-learn>=1.0.0",
     "pandas>=1.3.0",
     "networkx>=2.6.0",
-    "matplotlib",
+    "matplotlib>=3.5.0",
 ]
 readme = "README.md"
 dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://github.com/konstantingoe/mixed-gm"
+Issues = "https://github.com/konstantingoe/mixed-gm/issues"
+Documentation = "https://konstantingoe.github.io/mixed-gm/"
 
 [project.optional-dependencies]
 dev = [
@@ -38,7 +65,21 @@ dev = [
 ]
 
 [tool.hatch.version]
-path = "hume/__init__.py"
+path = "humpy/__init__.py"
+
+[tool.hatch.build.targets.wheel]
+packages = ["humpy"]
+
+[tool.hatch.build.targets.sdist]
+exclude = [
+    "venv_mixed-gm/",
+    "*.egg-info/",
+    "scribbling/",
+    ".ruff_cache/",
+    ".mypy_cache/",
+    ".pytest_cache/",
+    "site/",
+]
 
 [tool.ruff]
 preview = true
@@ -91,7 +132,7 @@ max-complexity = 10
 allow-multiline = false
 
 [tool.pytest.ini_options]
-testpaths = ["tests", "hume"]
+testpaths = ["tests", "humpy"]
 
 [tool.mypy]
 python_version = "3.10"
